feat: seperate configuration into mixins
The class Config is split into CommonConfigMixin, ServerConfigMixin and
WorkerConfigMixin. With these mixins, we can set options from args based on
the class attributes.
